delete contact for Supplier too

+++ _ressources/_inc/pages/scripts/form-validation.js
@@ -524,7 +524,7 @@ var FormValidation = function () {
 
                 percent: {
                     required: true,
-                    numeric: true
+                    float: true
                 },
             },
